[feat] OpenApiDTO 처리: WithdrawMoneyResponseDTO
