+++ src/main/kotlin/com/chatandpay/api/dto/OpenApiDTO.kt
@@ -63,4 +63,35 @@ class OpenApiDTO {
     )
 
 
+
+    @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
+    data class WithdrawMoneyResponseDTO(
+        val apiTranId: String,
+        val apiTranDtm: String,
+        val rspCode: String,
+        val rspMessage: String,
+        val dpsBankCodeStd: String,
+        val dpsBankCodeSub: String,
+        val dpsBankName: String,
+        val dpsAccountNumMasked: String,
+        val dpsPrintContent: String,
+        val dpsAccountHolderName: String,
+        val bankTranId: String,
+        val bankTranDate: String,
+        val bankCodeTran: String,
+        val bankRspCode: String,
+        val bankRspMessage: String,
+        val fintechUseNum: String,
+        val accountAlias: String,
+        val bankCodeStd: String,
+        val bankCodeSub: String,
+        val bankName: String,
+        val savingsBankName: String?,
+        val accountNumMasked: String,
+        val printContent: String,
+        val accountHolderName: String,
+        val tranAmt: String,
+        val wdLimitRemainAmt: String
+    )
+
 }
